backend/code/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
gunicorn==20.0.4
mysqlclient==1.4.6
ptvsd==4.3.2
pylint==2.4.4

